Restructured for better use // Finalized documentation.
