+++ lib/ci-github-bot.js
@@ -35,7 +35,7 @@ class CiGithubBot {
      * @param {ConfigurationComment} configuration Instance of comment configuration (template,
      *   variables ...)
      */
-    createPullRequestComment(configuration) {
+    createComment(configuration) {
         if (null === this.getConfiguration().getPullRequestNumber()) {
             throw new Error('Failed creating pull request comment: You need to configure the number of' +
                 ' the pull request before creating a comment.');
