+++ lib/circleci-github-bot.d.ts
