+++ lib/circleci-github-bot.js
