+++ lib/configuration/comment.d.ts
