+++ lib/configuration/comment.js
