+++ lib/configuration/github.d.ts
