+++ lib/configuration/github.js
