+++ lib/renderer/vanilla-js.d.ts
