+++ lib/renderer/vanilla-js.js
