+++ tsconfig.json
@@ -32,7 +32,7 @@
   "exclude": [
     "./node_modules",
     "./demo",
-    "./dist",
+    "./lib",
     "./docs"
   ]
 }
